Gen AI Projects/AI Resume-Job Insight/Backend/requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.1
starlette==0.37.2

# Streamlit (backend wrapper)
streamlit==1.38.0

# LangChain + Google Generative AI
langchain==0.2.12
langchain-core==0.2.23
langchain-google-genai==1.0.6

# PDF handling
PyPDF2==3.0.1
fpdf==1.7.2

# Environment variables
python-dotenv==1.0.1

# Utilities
requests==2.32.3
